Incorporate BINA suggestion
Towards v01-08

The next release of the model will be a major overhaul with several important revisions.
These revision are the results of several months of evaluation, feedback collection and correction of v01.07 by the members of the Hardware and Settings sub-group of the BINA Quality Control and Data Management Working Group.

Most changes will be added in one major update.

During this transition period, the .graphml, .jpg, and .pdf artifacts will be updated on a regular basis while the .xsd and the .xls files will be changed in one major push.

However the following few changes have been added already to the .xsd

-- Change description of Manufacturer Spec -->
This is the base from which all microscope hardware components (e.g., Objective, Filter etc.) are extended.  It provides attributes for recording common properties of these components such as Manufacturer name, Model, Serial Number etc.

-- Added AcquisitionSoftware element